prefactor: extract EncodingReader into its own file
note that some internal classes changed:

- Nokogiri::HTML4::Document::EncodingReader → Nokogiri::HTML4::EncodingReader
- Nokogiri::HTML4::Document::EncodingFound → Nokogiri::HTML4::EncodingReader::EncodingFound
